REFACTOR: re-package composite key class
