+++ src/main/java/kr/texturized/muus/infrastructure/repository/ImageRepository.java
@@ -1,12 +1,12 @@
 package kr.texturized.muus.infrastructure.repository;
 
 import kr.texturized.muus.domain.entity.Image;
-import kr.texturized.muus.domain.entity.ImageId;
+import kr.texturized.muus.domain.entity.fk.ImageFk;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
  * JPA Repository for Image.
  */
-public interface ImageRepository extends JpaRepository<Image, ImageId> {
+public interface ImageRepository extends JpaRepository<Image, ImageFk> {
 
 }
