+++ src/main/java/kr/texturized/muus/infrastructure/repository/KeywordRepository.java
@@ -1,12 +1,12 @@
 package kr.texturized.muus.infrastructure.repository;
 
 import kr.texturized.muus.domain.entity.Keyword;
-import kr.texturized.muus.domain.entity.KeywordId;
+import kr.texturized.muus.domain.entity.fk.KeywordFk;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 /**
  * JPA repository for Keyword.
  */
-public interface KeywordRepository extends JpaRepository<Keyword, KeywordId> {
+public interface KeywordRepository extends JpaRepository<Keyword, KeywordFk> {
 
 }
